Fazer a lógica do usuário só conseguir excluir e editar a promoção que postou, adicionando classes de testes com JUnit

+++ API/src/main/java/promoNice/API/model/UsuarioModel.java
@@ -12,7 +12,7 @@ public class UsuarioModel {
     @GeneratedValue(strategy = GenerationType.IDENTITY) // Isso fará com que o banco gere o valor automaticamente
     @Column(name = "id")
 
-    public Integer id;
+    public Long id;
 
     @Column (name = "nome")
     public String nome;
@@ -26,11 +26,11 @@ public class UsuarioModel {
     @Column  (name = "data_criacao")
     public LocalDateTime dataCriacao;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
